# pkg/handler/docs/swagger.yaml
basePath: /
definitions:
...
  handler.ApiResponse-handler_mfaConfirmResponse:
    properties:
      data:
        $ref: '#/definitions/handler.mfaConfirmResponse'
      error:
        type: string
    type: object
  handler.ApiResponse-map_string_string:
    properties:
      data:
        $ref: '#/definitions/map_string_string'
      error:
        type: string
    type: object
...
  handler.ApiResponse-service_MFAEnrollResponse:
    properties:
      data:
        $ref: '#/definitions/service.MFAEnrollResponse'
      error:
        type: string
    type: object
  handler.ApiResponse-service_TokenResponse:
    properties:
      data:
        $ref: '#/definitions/service.TokenResponse'
      error:
        type: string
    type: object
  handler.ApiResponse-string:
    properties:
      data:
        type: string
      error:
        type: string
    type: object
...
  handler.mfaConfirmRequest:
    properties:
      code:
        type: string
    type: object
  handler.mfaConfirmResponse:
    properties:
      recovery_codes:
        items:
          type: string
        type: array
    type: object
  handler.mfaLoginVerifyRequest:
    properties:
      code:
        type: string
      mfa_token:
        type: string
    type: object
  map_string_string:
    additionalProperties:
      type: string
    type: object
...
  service.MFAEnrollResponse:
    properties:
      otpauth_url:
        type: string
      secret:
        type: string
    type: object
...
  service.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
info:
  contact: {}
  description: Authentication service/library for Go.
  title: EzAuth API
  version: "1.0"
paths:
  /auth/api/mfa/confirm:
    post:
      consumes:
      - application/json
      description: Validates a TOTP code and enables MFA, returning one-time recovery
        codes
      parameters:
      - description: MFA Confirm Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.mfaConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ApiResponse-handler_mfaConfirmResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ApiResponse-string'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ApiResponse-string'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Confirm TOTP MFA enrollment
      tags:
      - mfa
  /auth/api/mfa/disable:
    post:
      consumes:
      - application/json
      description: Validates a TOTP or recovery code and disables MFA
      parameters:
      - description: MFA Disable Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.mfaConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ApiResponse-map_string_string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ApiResponse-string'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ApiResponse-string'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Disable TOTP MFA
      tags:
      - mfa
  /auth/api/mfa/enroll:
    post:
      description: Generates a new TOTP secret for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ApiResponse-service_MFAEnrollResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ApiResponse-string'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ApiResponse-string'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Begin TOTP MFA enrollment
      tags:
      - mfa
  /auth/api/mfa/login/verify:
    post:
      consumes:
      - application/json
      description: Exchanges an mfa_token and TOTP/recovery code for session tokens
      parameters:
      - description: MFA Login Verify Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.mfaLoginVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ApiResponse-service_TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ApiResponse-string'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ApiResponse-string'
      security:
      - ApiKeyAuth: []
      summary: Complete MFA step-up login
      tags:
      - mfa
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"